Refactor GUI modules
